continue prune / uncache work
